+++ game/GameObjects/Enemy.cs
@@ -5,12 +5,12 @@ namespace Blok3Game.GameObjects
 {
     public class Enemy : Character
     {
-        public int EnemyMoveSpeed;
+        public float EnemyMoveSpeed;
         public Vector2 steering;
         public Vector2 desired_velocity;
         public int EnemyShootCooldown = 120;
 
-        public Enemy(int hitPoints, int moveSpeed, Vector2 position, string assetName, int layer = 0, string id = "", int sheetIndex = 0) : base(hitPoints, moveSpeed, position, assetName)
+        public Enemy(int hitPoints, float moveSpeed, Vector2 position, string assetName, int layer = 0, string id = "", int sheetIndex = 0) : base(hitPoints, moveSpeed, position, assetName)
         {
             EnemyMoveSpeed = moveSpeed;
         }
