+++ game/GameObjects/ShootingEnemy.cs
@@ -4,7 +4,7 @@ using Blok3Game.GameObjects;
 public class ShootingEnemy : Enemy
 {
     public int EnemyHitPoints;
-    public ShootingEnemy(int hitPoints, int moveSpeed, Vector2 position) : base(hitPoints, moveSpeed, position, "Images/Characters/rat", 0, " ", 0)
+    public ShootingEnemy(int hitPoints, float moveSpeed, Vector2 position) : base(hitPoints, moveSpeed, position, "Images/Characters/rat", 0, " ", 0)
     {
         EnemyHitPoints = hitPoints;
     }
